+++ Source/EventFlow.TestHelpers/WaitHelper.cs
@@ -1,19 +1,19 @@
 ﻿// The MIT License (MIT)
-// 
+//
 // Copyright (c) 2015-2016 Rasmus Mikkelsen
 // Copyright (c) 2015-2016 eBay Software Foundation
 // https://github.com/rasmus/EventFlow
-// 
+//
 // Permission is hereby granted, free of charge, to any person obtaining a copy of
 // this software and associated documentation files (the "Software"), to deal in
 // the Software without restriction, including without limitation the rights to
 // use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of
 // the Software, and to permit persons to whom the Software is furnished to do so,
 // subject to the following conditions:
-// 
+//
 // The above copyright notice and this permission notice shall be included in all
 // copies or substantial portions of the Software.
-// 
+//
 // THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
 // IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS
 // FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR
@@ -23,14 +23,23 @@
 //
 
 using System;
+using System.Threading.Tasks;
 
-namespace EventFlow.Configuration
+namespace EventFlow.TestHelpers
 {
-    public interface IEventFlowConfiguration
+    public static class WaitHelper
     {
-        int PopulateReadModelEventPageSize { get; }
-        int NumberOfRetriesOnOptimisticConcurrencyExceptions { get; }
-        TimeSpan DelayBeforeRetryOnOptimisticConcurrencyExceptions { get; }
-        bool ThrowSubscriberExceptions { get; }
+        public static async Task WaitAsync(TimeSpan timeout, Func<Task<bool>> waitFor)
+        {
+            var start = DateTimeOffset.Now;
+            while (start + timeout > DateTimeOffset.Now)
+            {
+                var result = await waitFor().ConfigureAwait(false);
+                if (result) return;
+                await Task.Delay(TimeSpan.FromSeconds(0.25)).ConfigureAwait(false);
+            }
+
+            throw new ApplicationException("Failed to wait");
+        }
     }
 }